+++ data/src/main/java/ddd/buyornot/data/repository/login/AuthRepository.kt
@@ -17,5 +17,5 @@ interface AuthRepository {
 
     suspend fun logout(): Result<Unit>
 
-    suspend fun logoutRemote(): Result<BaseApiResponse<String>>
+    suspend fun logoutRemote(token: String): Result<BaseApiResponse<String>>
 }
