+++ data/src/main/java/ddd/buyornot/data/repository/login/AuthRepositoryImpl.kt
@@ -35,7 +35,7 @@ class AuthRepositoryImpl @Inject constructor(
         return authLocalDataSource.clearLocalData()
     }
 
-    override suspend fun logoutRemote(): Result<BaseApiResponse<String>> {
-        return authRemoteDataSource.postLogout()
+    override suspend fun logoutRemote(token: String): Result<BaseApiResponse<String>> {
+        return authRemoteDataSource.postLogout(token)
     }
 }
