+++ data/src/main/java/ddd/buyornot/data/source/login/AuthRemoteDataSourceImpl.kt
@@ -17,7 +17,7 @@ class AuthRemoteDataSourceImpl @Inject constructor(
         loginService.postKakaoAuth(KaKaoAuthRequest(token))
     }
 
-    override suspend fun postLogout(): Result<BaseApiResponse<String>> = runCatching {
-        loginService.postLogout()
+    override suspend fun postLogout(token: String): Result<BaseApiResponse<String>> = runCatching {
+        loginService.postLogout(KaKaoAuthRequest(token))
     }
 }
